requirements.txt
six
requests
pytz
python-jenkins
jinja2
